rename namespace Media to File
This is a bit more generic and allows us to have additional File
abstractions in here later.
